refactor: move and rename Logo and SideBarSignpost
