+++ src/components/AppComponents/AppHeaderMenu/index.ts
@@ -1,3 +1,3 @@
 export { default as AppHeaderMenu } from "./AppHeaderMenu";
-export type { TransitionProps } from "./AppHeaderMenu";
+export type { TransitionProps, MenuConfig } from "./AppHeaderMenu";
 export { SideMenu } from "./AppHeaderMenu";
