+++ src/components/AppComponents/AppHeaderMenuSidebarSignpost/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./AppHeaderMenuSidebarSignpost";
