+++ src/components/AppComponents/LayoutSiteFooter/LayoutSiteFooter.tsx
@@ -1,7 +1,7 @@
 import { FC, ReactNode } from "react";
 import { useRouter } from "next/router";
 
-import Logo from "@/components/Logo";
+import Logo from "@/components/AppComponents/Logo";
 import OakLink from "@/components/OakLink";
 import { OAK_SOCIALS } from "@/components/SharedComponents/SocialButtons/SocialButtons";
 import LayoutSiteFooterSignpost from "@/components/AppComponents/LayoutSiteFooterSignpost";
